Fix bug: Chg PW => Back => Chg PW, blank screen
Fix a bug that when you click on Change Password, click Back, then
click on Change Password again, you get a blank screen.

This was caused because I was sharing a method that removed the form in
order to hide it on the screen and in the DOM.  However, this only
worked for handlebars - for forms in index.html, the form was
permanently moved.  Change Password form was in index.html.

assets/scripts/app.js
- changed listener for change password to a dedicated listener

assets/scripts/calendars/ui.js
- created dedicated method for going back to calendars from change
password

index.html
- changed id of change password button

Future: move to handlebar